requirements.txt
asgiref==3.7.2
cloudinary==1.36.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==4.1
django-allauth==0.57.2
django-crispy-forms==2.0
django-heroku==0.3.1
django-summernote==0.8.20.0
environ==1.0
gunicorn==20.1.0
oauthlib==3.2.2
psycopg2==2.9.9
PyJWT==2.8.0
python3-openid==3.2.0
requests-oauthlib==1.4.0
sqlparse==0.4.4
urllib3==1.26.18
whitenoise==6.5.0
